messeges minimal done

+++ transport/sanic/endpoints/users/user.py
@@ -6,8 +6,9 @@ from api.response import ResponseUserDto
 from db.database import DBSession
 from db.exceptions import DBUserNotExistsException, DBDataException, DBIntegrityException
 from db.queries import user as user_queries
+from helpers.auth import read_token, get_id_from_token
 from transport.sanic.endpoints import BaseEndpoint
-from transport.sanic.exceptions import SanicUserNotFound, SanicDBException
+from transport.sanic.exceptions import SanicUserNotFound, SanicDBException, SanicForbidden
 
 
 class UserEndpoint(BaseEndpoint):
@@ -17,7 +18,7 @@ class UserEndpoint(BaseEndpoint):
         try:
             message = user_queries.get_user(session, uid=uid)
         except DBUserNotExistsException:
-            raise SanicUserNotFound('Message not found')  # change
+            raise SanicUserNotFound('User not found')  # change
 
         response_model = ResponseUserDto(message)
 
@@ -26,6 +27,9 @@ class UserEndpoint(BaseEndpoint):
     async def method_patch(
             self, request: Request, body: dict, session: DBSession, uid: int, *args, **kwargs
     ) -> BaseHTTPResponse:
+
+        if uid != get_id_from_token(request):
+            raise SanicForbidden('You have no rights to edit this user')
 
         request_model = RequestPatchUserDto(body)
 
@@ -46,6 +50,9 @@ class UserEndpoint(BaseEndpoint):
     async def method_delete(
             self, request: Request, body: dict, session: DBSession, uid: int, *args, **kwargs
     ) -> BaseHTTPResponse:
+
+        if uid != get_id_from_token(request):
+            raise SanicForbidden('You have no rights to delete this user')
 
         try:
             user = user_queries.delete_user(session, uid)
